dependencies/auth/auth.py
from functools import wraps
from flask import request, jsonify
import jwt

import os
from dotenv import load_dotenv

load_dotenv()

SECRET_KEY = os.getenv('JWT_SECRET_KEY')

def token_required(role):
    def decorator(f):
        @wraps(f)
        def decorated(*args, **kwargs):
            token = request.headers.get('Authorization')

            if not token:
                return jsonify({'message': 'Token is missing'}), 401

            try:
                data = jwt.decode(token, SECRET_KEY, algorithms=['HS256'])
                user_role = data.get('role')
                if user_role != role:
                    return jsonify({'message': 'Unauthorized'}), 403
            except jwt.ExpiredSignatureError:
                return jsonify({'message': 'Token has expired'}), 401
            except jwt.InvalidTokenError:
                return jsonify({'message': 'Invalid token'}), 401

            return f(*args, **kwargs)

        return decorated
    return decorator
